mcp-servers/yahoo-finance/requirements.txt
fastmcp>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
azure-identity>=1.17.0
azure-keyvault-secrets>=4.8.0
yfinance>=0.2.40
